feat(common): add JSON schema generators and validators
