+++ packages/ceramic-common/.gitignore
@@ -1,3 +1,4 @@
 node_modules/
 coverage/
 lib/
+src/generated/
