+++ packages/ceramic-common/package.json
@@ -22,10 +22,11 @@
     "prepublishOnly": "npm run build",
     "prebuild": "npm run clean",
     "lint": "./node_modules/.bin/eslint ./src --ext .js,.jsx,.ts,.tsx",
-    "clean": "rm -rf ./lib"
+    "clean": "rm -rf ./lib",
+    "generate": "cd scripts; ./generate_schemas.sh"
   },
   "dependencies": {
-    "@types/lodash.clonedeep": "^4.5.6",
+    "ajv": "^6.12.3",
     "did-resolver": "^2.0.1"
   },
   "devDependencies": {
@@ -35,8 +36,10 @@
     "@babel/preset-typescript": "^7.9.0",
     "@types/encoding-down": "^5.0.0",
     "@types/events": "^3.0.0",
-    "@types/node": "^13.9.5",
+    "@types/lodash.clonedeep": "^4.5.6",
+    "@types/node": "^13.13.14",
     "@types/node-fetch": "^2.5.5",
+    "@types/prettier": "^2.0.2",
     "@typescript-eslint/eslint-plugin": "^2.19.0",
     "@typescript-eslint/parser": "^2.19.0",
     "babel-jest": "^25.1.0",
@@ -46,7 +49,11 @@
     "ipfs": "^0.46.0",
     "jest": "^25.1.0",
     "js-sha256": "^0.9.0",
-    "tmp-promise": "^2.0.2"
+    "json-schema-to-typescript": "^9.1.1",
+    "tmp-promise": "^2.0.2",
+    "ts-node": "^8.10.2",
+    "tsc-watch": "^4.2.9",
+    "typescript-json-schema": "^0.42.0"
   },
   "jest": {
     "transformIgnorePatterns": [
