+++ packages/ceramic-common/schemas/string-map.json
@@ -1,0 +1,22 @@
+{
+    "$schema": "http://json-schema.org/draft-07/schema#",
+    "definitions": {
+        "StringMap": {
+            "additionalProperties": {
+            },
+            "properties": {
+                "additionalProperties": {
+                    "type": "string"
+                },
+                "required": {
+                },
+                "title": {
+                },
+                "type": {
+                }
+            },
+            "type": "object"
+        }
+    }
+}
+
