+++ packages/ceramic-common/src/index.ts
@@ -3,3 +3,4 @@ export * from './ceramic-api'
 export * from './context'
 export * from './doctype'
 export * from './user'
+export * from './utils/doctype-utils'
